+++ Recipes/RecipeDetailCollectionViewCell.swift
@@ -9,44 +9,6 @@
 import UIKit
 import QuartzCore
 
-extension RecipeDetailCollectionViewCell: UIScrollViewDelegate {
-  func scrollViewDidScroll(scrollView: UIScrollView) {
-    self.scrollViewDidScroll?(contentOffset: scrollView.contentOffset)
-  }
-}
-
-extension RecipeDetailCollectionViewCell: UITextViewDelegate {
-  func textView(textView: UITextView, shouldChangeTextInRange range: NSRange, replacementText text: String) -> Bool {
-    if range.length <= 1 && text.isEmpty && textView.text.isEmpty {
-      self.removeInstructionView(withInstructionTextView: textView, shiftFocusToPrevious: true)
-      return false
-    }
-    if text == "\n" {
-      if textView.text.isEmpty {
-        return false
-      }
-      if let instructionView = textView.superview?.superview?.superview as? RecipeInstructionView {
-        if var index = self.instructionsStackView.arrangedSubviews.indexOf(instructionView) {
-          index = index + 1
-          if index < self.instructionsStackView.arrangedSubviews.count {
-            self.addInstructionView("", focusOnTextView: true, atIndex: index)
-          } else {
-            self.addInstructionView("", focusOnTextView: true, atIndex: nil)
-          }
-        }
-      }
-      return false
-    }
-    return true
-  }
-  
-  func textViewDidEndEditing(textView: UITextView) {
-    if textView.text.isEmpty {
-      self.removeInstructionView(withInstructionTextView: textView, shiftFocusToPrevious: false)
-    }
-  }
-}
-
 class RecipeDetailCollectionViewCell: UICollectionViewCell {
   @IBOutlet var scrollView: UIScrollView!
   @IBOutlet var nameTextField: UITextField!
@@ -90,6 +52,7 @@ class RecipeDetailCollectionViewCell: UICollectionViewCell {
             if let image = image {
               NSOperationQueue.mainQueue().addOperationWithBlock({
                 self.photoImageView.image = image
+                self.backgroundImageView.image = image
               })
             }
           })
