+++ Recipes/RecipesFilterViewController+UITableViewDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  RecipesFilterViewController+UITableViewDelegate.swift
+//  Recipes
+//
+//  Created by Nofel Mahmood on 25/10/2015.
+//  Copyright © 2015 Hyper. All rights reserved.
+//
+
+import UIKit
+
+extension RecipesFilterViewController: UITableViewDelegate {
+  func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+    self.delegate?.recipesFilterViewController(self, didSelectFilter: self.selectedRecipesFilter())
+  }
+}
