+++ Recipes/RecipesListViewController+UINavigationControllerDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  RecipesListViewController+UINavigationControllerDelegate.swift
+//  Recipes
+//
+//  Created by Nofel Mahmood on 25/10/2015.
+//  Copyright © 2015 Hyper. All rights reserved.
+//
+
+import UIKit
+
+extension RecipesListViewController: UINavigationControllerDelegate {
+  func navigationController(navigationController: UINavigationController, animationControllerForOperation operation: UINavigationControllerOperation, fromViewController fromVC: UIViewController, toViewController toVC: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+    return self.toRecipeDetailViewControllerTransition
+  }
+  
+  //  func navigationController(navigationController: UINavigationController, interactionControllerForAnimationController animationController: UIViewControllerAnimatedTransitioning) -> UIViewControllerInteractiveTransitioning? {
+  //    return self.interactivePopTransition
+  //  }
+}
