+++ Recipes/RecipesListViewController+UIPopoverPresentationControllerDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  RecipesListViewController+UIPopoverPresentationControllerDelegate.swift
+//  Recipes
+//
+//  Created by Nofel Mahmood on 25/10/2015.
+//  Copyright © 2015 Hyper. All rights reserved.
+//
+
+import UIKit
+
+extension RecipesListViewController: UIPopoverPresentationControllerDelegate {
+  func adaptivePresentationStyleForPresentationController(controller: UIPresentationController) -> UIModalPresentationStyle {
+    return UIModalPresentationStyle.None
+  }
+}
