+++ Recipes/RecipesListViewController+UIViewControllerTransitioningDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  RecipesListViewController+UIViewControllerTransitioningDelegate.swift
+//  Recipes
+//
+//  Created by Nofel Mahmood on 25/10/2015.
+//  Copyright © 2015 Hyper. All rights reserved.
+//
+
+import UIKit
+
+extension RecipesListViewController: UIViewControllerTransitioningDelegate {
+  func animationControllerForPresentedController(presented: UIViewController, presentingController presenting: UIViewController, sourceController source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+    return self.toRecipesSearchViewControllerTransition
+  }
+  
+  func animationControllerForDismissedController(dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+    return self.toRecipesSearchViewControllerTransition
+  }
+}
