+++ Recipes/RecipesSearchViewController.swift
@@ -14,73 +14,11 @@ enum RecipeSearchSegue: String {
 
 let RecipeSearchTableViewCellIdentifier = "RecipeSearchTableViewCell"
 
-// MARK: UISearchControllerDelegate
-extension RecipesSearchViewController: UISearchControllerDelegate {
-  
-  func didDismissSearchController(searchController: UISearchController) {
-    self.presentingViewController?.dismissViewControllerAnimated(true, completion: nil)
-  }
-}
-
-// MARK: UISearchResultsUpdating
-extension RecipesSearchViewController: UISearchResultsUpdating {
-  func updateSearchResultsForSearchController(searchController: UISearchController) {
-    guard let searchString = searchController.searchBar.text else {
-      self.searchResults.removeAll()
-      self.tableView.reloadData()
-      return
-    }
-    guard let recipes = self.recipes else {
-      return
-    }
-    self.searchResults = recipes.filter({ (recipe) -> Bool in
-      if recipe.name.rangeOfString(searchString, options: NSStringCompareOptions.CaseInsensitiveSearch, range: nil, locale: nil) != nil {
-        return true
-      }
-      return false
-    })
-    self.tableView.reloadData()
-  }
-}
-
-// MARK: UITableViewDataSource
-extension RecipesSearchViewController: UITableViewDataSource {
-  func numberOfSectionsInTableView(tableView: UITableView) -> Int {
-    return 1
-  }
-  
-  func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-    return self.searchResults.count
-  }
-  
-  func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-    let cell = tableView.dequeueReusableCellWithIdentifier(RecipeSearchTableViewCellIdentifier, forIndexPath: indexPath) as! RecipeSearchTableViewCell
-    let recipe = self.searchResults[indexPath.row]
-    cell.nameLabel.text = recipe.name
-    recipe.photo({ photo in
-      if let photo = photo, let index = self.searchResults.indexOf(recipe) {
-        NSOperationQueue.mainQueue().addOperationWithBlock({
-          if let correspondingCell = self.tableView.cellForRowAtIndexPath(NSIndexPath(forRow: index, inSection: 0)) as? RecipeSearchTableViewCell {
-            correspondingCell.photoImageView.image = photo
-          }
-        })
-      }
-    })
-    return cell
-  }
-}
-
-// MARK: UITableViewDelegate
-extension RecipesSearchViewController: UITableViewDelegate {
-  func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-    self.performSegueWithIdentifier(RecipeSearchSegue.RecipeModifier.rawValue, sender: self)
-  }
-}
-
-// MARK: -
 class RecipesSearchViewController: UIViewController {
   
   @IBOutlet var tableView: UITableView!
+  @IBOutlet var backgroundImageView: UIImageView!
+  var imageToBlur: UIImage!
   let searchController = UISearchController(searchResultsController: nil)
   
   var recipes: [RecipeViewModel]!
@@ -103,6 +41,7 @@ class RecipesSearchViewController: UIViewController {
   
   override func viewDidAppear(animated: Bool) {
     super.viewDidAppear(animated)
+    self.backgroundImageView.image = self.imageToBlur
     self.performSelector("focusSearchBar", withObject: nil, afterDelay: 0.1)
   }
